libstagefright: Increase raw decoder buffers to 8
For WAV clips with wrong header values, the first
attempt at PCM offloading is unsuccessful. This leads
to the creation of a  raw decoder and another attempt
at offloading is tried with decoded PCM offload.
This attempt succeeds, and the graph now has a GenericSource,
NuPlayerDecoder and NuPlayerRenderer in it. Raw decoder
has 4 output buffers, but this is not enough to fill the
data at the rate NuPlayerRenderer needs for offloading.
Fix this by increasing the buffer count for raw decoder.

Change-Id: I29acc6b35fa9d5ab94e1df551f4cd89685527162